requirements.txt
langchain
langchain-community
langgraph
ollama
faiss-cpu
pypdf
beautifulsoup4
requests
streamlit
chainlit